+++ src/common/utxobased/info/digibyte.ts
@@ -40,7 +40,7 @@ const engineInfo: EngineInfo = {
   serverConfigs: [
     {
       type: 'blockbook-nownode',
-      uris: ['dgbbook.nownodes.io']
+      uris: ['https://dgbbook.nownodes.io']
     }
   ],
   formats: ['bip49', 'bip84', 'bip44', 'bip32'],