+++ src/common/utxobased/info/groestlcoin.ts
@@ -46,7 +46,7 @@ const engineInfo: EngineInfo = {
   serverConfigs: [
     {
       type: 'blockbook-nownode',
-      uris: ['grsbook.nownodes.io']
+      uris: ['https://grsbook.nownodes.io']
     }
   ],
   formats: ['bip49', 'bip84', 'bip44', 'bip32'],